requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
autoevals==0.0.119
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.2.0
braintrust_core==0.0.58
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
cloudpathlib==0.20.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.3.2
debugpy==1.8.12
decorator==5.1.1
deepeval==2.4.3
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.2.0
faiss-cpu==1.10.0
fastjsonschema==2.21.1
filelock==3.17.0
filetype==1.2.0
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
galileo==0.0.4
galileo-core==3.9.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.39
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langcodes==3.5.0
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.8
language_data==1.3.0
Levenshtein==0.26.1
llama-cloud==0.1.13
llama-cloud-services==0.6.1
llama-index==0.12.19
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.0
llama-index-core==0.12.19
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.8
llama-index-llms-openai==0.3.20
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.27.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.61.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
peewee==3.17.9
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
portalocker==3.1.1
preshed==3.0.9
prompt_toolkit==3.0.50
promptquality==1.3.1
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pypdf==5.3.0
pytest==8.3.4
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytickersymbols==1.13.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentry-sdk==2.22.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.42.2
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
widgetsnbextension==4.0.13
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.54
zipp==3.21.0
zstandard==0.23.0
